Ajustes de interface e otimizando imports

+++ app/src/main/java/com/neuralnet/maisfinancas/ui/components/ItemDespesa.kt
@@ -15,6 +15,7 @@ import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.unit.dp
 import com.neuralnet.maisfinancas.model.Despesa
 import com.neuralnet.maisfinancas.util.formattedDate
+import com.neuralnet.maisfinancas.util.toReal
 import java.text.SimpleDateFormat
 import java.util.Date
 import java.util.Locale
@@ -29,7 +30,7 @@ fun ItemDespesa(despesa: Despesa, modifier: Modifier = Modifier) {
                 Text(text = despesa.data.formattedDate(), fontWeight = FontWeight.Light)
             }
 
-            Text(text = despesa.valor.toDouble().toReal(), fontStyle = MaterialTheme.typography.bodyMedium.fontStyle, fontWeight = FontWeight.Medium)
+            Text(text = despesa.valor.toReal(), fontStyle = MaterialTheme.typography.bodyMedium.fontStyle, fontWeight = FontWeight.Medium)
 
         }
     }
